removed unneccesary settings
