--- Helpers/MIN.Helpers.Contracts/Models/Enums/SearchMethod.cs
@@ -1,17 +1,0 @@
-﻿namespace MIN.Helpers.Contracts.Models.Enums;
-
-/// <summary>
-/// Метод поиска комнаты
-/// </summary>
-public enum SearchMethod
-{
-    /// <summary>
-    /// Обход по кабинету
-    /// </summary>
-    ClassRoom = 0,
-
-    /// <summary>
-    /// Избранные комнаты
-    /// </summary>
-    Preferred = 1,
-}
