+++ Helpers/MIN.Helpers.Contracts/Models/Settings.cs
@@ -1,6 +1,4 @@
-﻿using MIN.Helpers.Contracts.Models.Enums;
-
-namespace MIN.Helpers.Contracts.Models;
+﻿namespace MIN.Helpers.Contracts.Models;
 
 /// <summary>
 /// Настройки
@@ -16,16 +14,6 @@ public class Settings
     /// Время ожидания поиска комнаты
     /// </summary>
     public int DiscoveryTimeout { get; set; } = 1500;
-
-    /// <summary>
-    /// Метод поиска комнат
-    /// </summary>
-    public SearchMethod SearchMethod { get; set; } = SearchMethod.ClassRoom;
-
-    /// <summary>
-    /// Избранные компьютеры
-    /// </summary>
-    public IEnumerable<string> PreferredPCNames { get; set; } = [];
 
     /// <summary>
     /// Порт для обнаружения в сети
